backend/requirements.txt
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asttokens==3.0.1
attrs==25.4.0
backports.zstd==1.2.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
bleach==6.3.0
blis==1.3.3
brotli==1.2.0
cachetools==6.2.2
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
colorlog==6.10.1
comm==0.2.3
confection==0.1.5
contourpy==1.3.3
coverage==7.13.1
cryptography==46.0.3
cycler==0.12.1
cymem==2.0.13
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
ecdsa==0.19.1
executing==2.2.1
fastapi==0.121.3
fastjsonschema==2.21.2
filelock==3.20.1
fonttools==4.61.1
fsspec==2025.12.0
fugashi==1.5.2
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-vision==3.11.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
inflate64==1.0.4
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jaconv==0.4.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jreadability==1.1.5
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.9
Levenshtein==0.27.3
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mistune==3.2.0
mpmath==1.3.0
multivolumefile==0.2.3
murmurhash==1.0.15
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.3.5
optuna==4.6.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
passlib==1.7.4
patool==4.0.3
pillow==12.0.0
platformdirs==4.5.1
pluggy==1.6.0
preshed==3.0.12
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.33.2
psutil==7.1.3
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybcj==1.0.7
pycparser==2.23
pycryptodomex==3.23.0
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
pyppmd==1.2.0
pysubs2==1.8.0
pytest==9.0.1
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-Levenshtein==0.27.3
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
pyzstd==0.19.1
RapidFuzz==3.14.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
sentence-transformers==5.2.0
setuptools==80.9.0
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
soupsieve==2.8.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
sqlmodel==0.0.27
srsly==2.5.2
stack-data==0.6.3
starlette==0.50.0
SudachiDict-core==20251022
SudachiPy==0.6.10
sympy==1.14.0
texttable==1.7.0
thefuzz==0.22.1
thinc==8.3.10
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.1
torch==2.9.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
typer-slim==0.20.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unidic-lite==1.0.8
urllib3==2.5.0
uvicorn==0.38.0
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.3
webencodings==0.5.1
wrapt==2.0.1
xgboost==3.1.2
